Save production charge in integer pence to avoid any floating point issues

+++ database/migrations/2019_07_03_200146_event_production_charge.php
@@ -14,7 +14,7 @@ class EventProductionCharge extends Migration
     public function up()
     {
         Schema::table('events', function(Blueprint $table) {
-            $table->decimal('production_charge', 8,2)
+            $table->integer('production_charge')
                   ->after('venue_type')
                   ->nullable();
         });
